+++ src/client/auth_handler.rs
@@ -1,12 +1,21 @@
+use std::io;
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use tokio::sync::Mutex;
+
 use crate::types::{
     AuthorizationStateWaitCode, AuthorizationStateWaitEncryptionKey,
     AuthorizationStateWaitOtherDeviceConfirmation, AuthorizationStateWaitPassword,
     AuthorizationStateWaitPhoneNumber, AuthorizationStateWaitRegistration,
 };
-use async_trait::async_trait;
-use std::io;
-use std::sync::Arc;
-use tokio::sync::Mutex;
+
+/// `ClientIdentifier` allows to determine if client is bot (with bot token as identifier) or client (with a phone number)
+#[derive(Debug, Clone)]
+pub enum ClientIdentifier {
+    PhoneNumber(String),
+    BotToken(String),
+}
 
 /// `AuthStateHandler` trait provides methods that returns data, required for authentication
 /// It allows you to handle particular "auth states", such as [WaitPassword](crate::types::AuthorizationStateWaitPassword), [WaitPhoneNumber](crate::types::AuthorizationStateWaitPhoneNumber) and so on.
@@ -31,11 +40,11 @@ pub trait AuthStateHandler {
     ) -> String;
     /// Returns password
     async fn handle_wait_password(&self, wait_password: &AuthorizationStateWaitPassword) -> String;
-    /// Returns phone number
-    async fn handle_wait_phone_number(
+    /// Returns [ClientIdentifier](crate::client::auth_handler::ClientIdentifier)
+    async fn handle_wait_client_identifier(
         &self,
         wait_phone_number: &AuthorizationStateWaitPhoneNumber,
-    ) -> String;
+    ) -> ClientIdentifier;
     /// Returns first_name and second_name
     async fn handle_wait_registration(
         &self,
@@ -59,7 +68,7 @@ impl ConsoleAuthStateHandler {
         Self
     }
 
-    fn wait_input() -> String {
+    fn wait_input(&self) -> String {
         let mut input = String::new();
         match io::stdin().read_line(&mut input) {
             Ok(_) => input.trim().to_string(),
@@ -72,7 +81,7 @@ impl ConsoleAuthStateHandler {
 impl AuthStateHandler for ConsoleAuthStateHandler {
     async fn handle_wait_code(&self, _wait_code: &AuthorizationStateWaitCode) -> String {
         println!("waiting for auth code");
-        ConsoleAuthStateHandler::wait_input()
+        self.wait_input()
     }
 
     async fn handle_encryption_key(
@@ -80,7 +89,7 @@ impl AuthStateHandler for ConsoleAuthStateHandler {
         _wait_encryption_key: &AuthorizationStateWaitEncryptionKey,
     ) -> String {
         println!("waiting for encryption key");
-        ConsoleAuthStateHandler::wait_input()
+        self.wait_input()
     }
 
     async fn handle_wait_password(
@@ -88,24 +97,39 @@ impl AuthStateHandler for ConsoleAuthStateHandler {
         _wait_password: &AuthorizationStateWaitPassword,
     ) -> String {
         println!("waiting for password");
-        ConsoleAuthStateHandler::wait_input()
+        self.wait_input()
     }
 
-    async fn handle_wait_phone_number(
+    async fn handle_wait_client_identifier(
         &self,
-        _wait_phone_number: &AuthorizationStateWaitPhoneNumber,
-    ) -> String {
-        println!("waiting for phone number");
-        ConsoleAuthStateHandler::wait_input()
+        _: &AuthorizationStateWaitPhoneNumber,
+    ) -> ClientIdentifier {
+        loop {
+            println!("choose one of phone number (p) or bot token (b)");
+            let inp = self.wait_input();
+            match inp.to_lowercase().trim() {
+                "b" => {
+                    println!("enter bot token");
+                    return ClientIdentifier::BotToken(self.wait_input());
+                }
+                "p" => {
+                    println!("enter phone number");
+                    return ClientIdentifier::PhoneNumber(self.wait_input());
+                }
+                _ => {
+                    // invalid input, next iteration}
+                    continue;
+                }
+            }
+        }
     }
-
     async fn handle_wait_registration(
         &self,
         _wait_registration: &AuthorizationStateWaitRegistration,
     ) -> (String, String) {
         loop {
             println!("waiting for first_name and second_name separated by comma");
-            let inp: String = ConsoleAuthStateHandler::wait_input();
+            let inp: String = self.wait_input();
             if let Some((f, l)) = split_string(inp, ',') {
                 return (f, l);
             }
@@ -151,9 +175,28 @@ impl AuthStateHandler for SignalAuthStateHandler {
         self.wait_signal().await
     }
 
-    async fn handle_wait_phone_number(&self, _: &AuthorizationStateWaitPhoneNumber) -> String {
-        log::info!("waiting for phone number");
-        self.wait_signal().await
+    async fn handle_wait_client_identifier(
+        &self,
+        _: &AuthorizationStateWaitPhoneNumber,
+    ) -> ClientIdentifier {
+        loop {
+            log::info!("choose one of phone number (p) or bot token (b)");
+            let inp = self.wait_signal().await;
+            match inp.to_lowercase().trim() {
+                "b" => {
+                    log::info!("enter bot token");
+                    return ClientIdentifier::BotToken(self.wait_signal().await);
+                }
+                "p" => {
+                    log::info!("enter phone number");
+                    return ClientIdentifier::PhoneNumber(self.wait_signal().await);
+                }
+                _ => {
+                    // invalid input, next iteration}
+                    continue;
+                }
+            }
+        }
     }
 
     async fn handle_wait_registration(